# Speed up BUFFER increases (minimize reallocs)